refactor(agent-mcp): collapse McpRegistryService into McpServerService; gather tools per request; switch to PinoLogger
