+++ src/agentMcp/agentMcp.module.ts
@@ -1,7 +1,6 @@
 import { Module } from '@nestjs/common'
 import { DiscoveryModule } from '@nestjs/core'
 import { ClerkModule } from '@/vendors/clerk/clerk.module'
-import { McpRegistryService } from './services/mcpRegistry.service'
 import { McpServerService } from './services/mcpServer.service'
 import { AgentMcpController } from './agentMcp.controller'
 import { AgentActorGuard } from './guards/AgentActor.guard'
@@ -11,12 +10,7 @@ import { AgentActorTokenService } from './services/agentActorToken.service'
 @Module({
   imports: [DiscoveryModule, ClerkModule],
   controllers: [AgentMcpController, AgentActorTokenController],
-  providers: [
-    McpRegistryService,
-    McpServerService,
-    AgentActorGuard,
-    AgentActorTokenService,
-  ],
-  exports: [McpRegistryService],
+  providers: [McpServerService, AgentActorGuard, AgentActorTokenService],
+  exports: [McpServerService],
 })
 export class AgentMcpModule {}
